requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
langgraph>=0.0.10
langchain-google-vertexai>=0.0.5
langchain-core>=0.1.10
PyGithub>=2.1.1
python-dotenv>=1.0.1
httpx>=0.26.0
ruff>=0.3.0
google-cloud-tasks>=2.16.0
